config/requirements.txt
Django==5.0.7
django-cleanup==8.1.0
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
pillow==10.4.0
PyJWT==2.9.0


